fix: FileChannel error even when file transfered
Sometimes the file channel was closed before the receiver was notified
of the success of the transfer. To fix that we removed the done event
listener on the receiver side so they will not close the channel. Only
the sender can close the channel now on file tranfer success.
Sometimes when closing the file channel un error can occur on the
receiver side, this error will be ignored if the file was successfully
tranfered.